requirements.txt
asgiref==3.11.1
coverage==7.13.5
Django==3.0.7
django-recurrence==1.11.1
gunicorn==25.3.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pyaml==26.2.1
python-dateutil==2.9.0.post0
python-decouple==3.3
pytz==2026.1.post1
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
whitenoise==6.12.0
